refactor(agents): trim embedded runner exports
