+++ src/agents/embedded-agent-runner-extraparams.live.test.ts
@@ -3,7 +3,7 @@ import type { Model } from "openclaw/plugin-sdk/llm";
 import { streamSimple } from "openclaw/plugin-sdk/llm";
 import { describe, expect, it } from "vitest";
 import type { OpenClawConfig } from "../config/config.js";
-import { applyExtraParamsToAgent } from "./embedded-agent-runner.js";
+import { applyExtraParamsToAgent } from "./embedded-agent-runner/extra-params.js";
 import { isLiveTestEnabled } from "./live-test-helpers.js";
 import { isLiveAuthDrift, isLiveBillingDrift } from "./live-test-provider-drift.js";
