+++ src/agents/embedded-agent-runner.splitsdktools.test.ts
@@ -1,10 +1,10 @@
 // Coverage for classifying SDK tools into the embedded runner runtime surface.
 import { describe, expect, it } from "vitest";
-import { splitSdkTools } from "./embedded-agent-runner.js";
 import {
   collectRegisteredToolNames,
   toSessionToolAllowlist,
 } from "./embedded-agent-runner/tool-name-allowlist.js";
+import { splitSdkTools } from "./embedded-agent-runner/tool-split.js";
 import { createStubTool } from "./test-helpers/agent-tool-stubs.js";
 
 describe("splitSdkTools", () => {
