+++ src/agents/embedded-agent-runner.ts
@@ -1,7 +1,6 @@
 // Embedded-agent runner barrel. Focused submodules own run orchestration,
 // compaction, queues, sandbox metadata, and SDK tool splitting.
 export { compactEmbeddedAgentSession } from "./embedded-agent-runner/compact.queued.js";
-export { applyExtraParamsToAgent } from "./embedded-agent-runner/extra-params.js";
 
 export { resolveEmbeddedSessionLane } from "./embedded-agent-runner/lanes.js";
 export { runEmbeddedAgent } from "./embedded-agent-runner/run.js";
@@ -18,8 +17,6 @@ export {
   resolveActiveEmbeddedRunSessionIdBySessionFile,
   waitForEmbeddedAgentRunEnd,
 } from "./embedded-agent-runner/runs.js";
-export { buildEmbeddedSandboxInfo } from "./embedded-agent-runner/sandbox-info.js";
-export { splitSdkTools } from "./embedded-agent-runner/tool-split.js";
 export type {
   EmbeddedAgentMeta,
   EmbeddedAgentCompactResult,
